+++ js/savecustomcss.php
@@ -29,6 +29,7 @@ $allowed = [
     '--selector-bg', '--blocktitle',
     '--text-title', '--text-status',
     '--font-small', '--font-large',
+    '--icon-font-size', '--icon-image-size',
 ];
 
 $sanitized = [];